// src/components/props/display.type.ts
export type Display =
  | 'flex'
  | 'flex-column'
  | 'flex-column-center'
  | 'flex-column-start'
  | 'flex-column-end'
  | 'flex-column-stretch'
  | 'flex-column-reverse'
  | 'flex-column-reverse-center'
  | 'flex-column-reverse-start'
  | 'flex-column-reverse-end'
  | 'flex-column-reverse-stretch'
  | 'flex-row'
  | 'flex-row-center'
  | 'flex-row-start'
  | 'flex-row-end'
  | 'flex-row-stretch'
  | 'flex-row-reverse'
  | 'flex-row-reverse-center'
  | 'flex-row-reverse-start'
  | 'flex-row-reverse-end'
  | 'flex-row-reverse-stretch'
  | 'inline-flex'
  | 'block'
  | 'inline-block'
  | 'inline-grid'
  | 'contents'
  | 'list-item'
  | 'revert'
  | 'table'
  | 'inline-table'
  | 'table-caption'
  | 'table-cell'
  | 'table-column'
  | 'table-column-group'
  | 'table-footer-group'
  | 'table-header-group'
  | 'table-row'
  | 'table-row-group'
  | 'none'
  | 'unset';